fix: harden db engine and orm setup

+++ src/reflexor/infra/db/models.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from sqlalchemy import JSON, Boolean, ForeignKey, Integer, String
+from sqlalchemy import JSON, BigInteger, Boolean, ForeignKey, Integer, String
 from sqlalchemy.dialects.postgresql import JSONB
 from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column
 
@@ -17,7 +17,7 @@ class EventRow(Base):
     event_id: Mapped[str] = mapped_column(String, primary_key=True)
     type: Mapped[str] = mapped_column(String, index=True)
     source: Mapped[str] = mapped_column(String)
-    received_at_ms: Mapped[int] = mapped_column(Integer)
+    received_at_ms: Mapped[int] = mapped_column(BigInteger)
     payload: Mapped[dict[str, object]] = mapped_column(JSON_VARIANT)
     dedupe_key: Mapped[str | None] = mapped_column(String, nullable=True)
 
@@ -28,9 +28,9 @@ class EventDedupeRow(Base):
     source: Mapped[str] = mapped_column(String, primary_key=True)
     dedupe_key: Mapped[str] = mapped_column(String, primary_key=True)
     event_id: Mapped[str] = mapped_column(String, ForeignKey("events.event_id"))
-    created_at_ms: Mapped[int] = mapped_column(Integer)
-    updated_at_ms: Mapped[int] = mapped_column(Integer)
-    expires_at_ms: Mapped[int] = mapped_column(Integer, index=True)
+    created_at_ms: Mapped[int] = mapped_column(BigInteger)
+    updated_at_ms: Mapped[int] = mapped_column(BigInteger)
+    expires_at_ms: Mapped[int] = mapped_column(BigInteger, index=True)
 
 
 class RunRow(Base):
@@ -38,9 +38,9 @@ class RunRow(Base):
 
     run_id: Mapped[str] = mapped_column(String, primary_key=True)
     parent_run_id: Mapped[str | None] = mapped_column(String, nullable=True)
-    created_at_ms: Mapped[int] = mapped_column(Integer, index=True)
-    started_at_ms: Mapped[int | None] = mapped_column(Integer, nullable=True)
-    completed_at_ms: Mapped[int | None] = mapped_column(Integer, nullable=True)
+    created_at_ms: Mapped[int] = mapped_column(BigInteger, index=True)
+    started_at_ms: Mapped[int | None] = mapped_column(BigInteger, nullable=True)
+    completed_at_ms: Mapped[int | None] = mapped_column(BigInteger, nullable=True)
 
 
 class ToolCallRow(Base):
@@ -52,9 +52,9 @@ class ToolCallRow(Base):
     permission_scope: Mapped[str] = mapped_column(String)
     idempotency_key: Mapped[str] = mapped_column(String, index=True)
     status: Mapped[str] = mapped_column(String)
-    created_at_ms: Mapped[int] = mapped_column(Integer)
-    started_at_ms: Mapped[int | None] = mapped_column(Integer, nullable=True)
-    completed_at_ms: Mapped[int | None] = mapped_column(Integer, nullable=True)
+    created_at_ms: Mapped[int] = mapped_column(BigInteger)
+    started_at_ms: Mapped[int | None] = mapped_column(BigInteger, nullable=True)
+    completed_at_ms: Mapped[int | None] = mapped_column(BigInteger, nullable=True)
     result_ref: Mapped[str | None] = mapped_column(String, nullable=True)
 
 
@@ -72,9 +72,9 @@ class TaskRow(Base):
     max_attempts: Mapped[int] = mapped_column(Integer)
     timeout_s: Mapped[int] = mapped_column(Integer)
     depends_on: Mapped[list[str]] = mapped_column(JSON_VARIANT)
-    created_at_ms: Mapped[int] = mapped_column(Integer)
-    started_at_ms: Mapped[int | None] = mapped_column(Integer, nullable=True)
-    completed_at_ms: Mapped[int | None] = mapped_column(Integer, nullable=True)
+    created_at_ms: Mapped[int] = mapped_column(BigInteger)
+    started_at_ms: Mapped[int | None] = mapped_column(BigInteger, nullable=True)
+    completed_at_ms: Mapped[int | None] = mapped_column(BigInteger, nullable=True)
     labels: Mapped[list[str]] = mapped_column(JSON_VARIANT)
     metadata_json: Mapped[dict[str, object]] = mapped_column("metadata", JSON_VARIANT)
 
@@ -87,8 +87,8 @@ class ApprovalRow(Base):
     task_id: Mapped[str] = mapped_column(String, ForeignKey("tasks.task_id"))
     tool_call_id: Mapped[str] = mapped_column(String, ForeignKey("tool_calls.tool_call_id"))
     status: Mapped[str] = mapped_column(String, index=True)
-    created_at_ms: Mapped[int] = mapped_column(Integer)
-    decided_at_ms: Mapped[int | None] = mapped_column(Integer, nullable=True)
+    created_at_ms: Mapped[int] = mapped_column(BigInteger)
+    decided_at_ms: Mapped[int | None] = mapped_column(BigInteger, nullable=True)
     decided_by: Mapped[str | None] = mapped_column(String, nullable=True)
     payload_hash: Mapped[str | None] = mapped_column(String, nullable=True)
     preview: Mapped[str | None] = mapped_column(String, nullable=True)
@@ -99,7 +99,7 @@ class RunPacketRow(Base):
 
     run_id: Mapped[str] = mapped_column(String, ForeignKey("runs.run_id"), primary_key=True)
     packet_version: Mapped[int] = mapped_column(Integer, default=1)
-    created_at_ms: Mapped[int] = mapped_column(Integer)
+    created_at_ms: Mapped[int] = mapped_column(BigInteger)
     packet: Mapped[dict[str, object]] = mapped_column(JSON_VARIANT)
 
 
@@ -110,18 +110,20 @@ class EventSuppressionRow(Base):
     event_type: Mapped[str] = mapped_column(String, index=True)
     event_source: Mapped[str] = mapped_column(String, index=True)
     signature: Mapped[dict[str, object]] = mapped_column(JSON_VARIANT)
-    window_start_ms: Mapped[int] = mapped_column(Integer)
+    window_start_ms: Mapped[int] = mapped_column(BigInteger)
     count: Mapped[int] = mapped_column(Integer)
     threshold: Mapped[int] = mapped_column(Integer)
     window_ms: Mapped[int] = mapped_column(Integer)
-    suppressed_until_ms: Mapped[int | None] = mapped_column(Integer, nullable=True, index=True)
+    suppressed_until_ms: Mapped[int | None] = mapped_column(
+        BigInteger, nullable=True, index=True
+    )
     resume_required: Mapped[bool] = mapped_column(Boolean)
-    cleared_at_ms: Mapped[int | None] = mapped_column(Integer, nullable=True)
+    cleared_at_ms: Mapped[int | None] = mapped_column(BigInteger, nullable=True)
     cleared_by: Mapped[str | None] = mapped_column(String, nullable=True)
     cleared_request_id: Mapped[str | None] = mapped_column(String, nullable=True)
-    created_at_ms: Mapped[int] = mapped_column(Integer)
-    updated_at_ms: Mapped[int] = mapped_column(Integer)
-    expires_at_ms: Mapped[int] = mapped_column(Integer, index=True)
+    created_at_ms: Mapped[int] = mapped_column(BigInteger)
+    updated_at_ms: Mapped[int] = mapped_column(BigInteger)
+    expires_at_ms: Mapped[int] = mapped_column(BigInteger, index=True)
 
 
 class IdempotencyLedgerRow(Base):
@@ -131,9 +133,9 @@ class IdempotencyLedgerRow(Base):
     tool_name: Mapped[str] = mapped_column(String, index=True)
     status: Mapped[str] = mapped_column(String, index=True)
     result_json: Mapped[dict[str, object]] = mapped_column(JSON_VARIANT)
-    created_at_ms: Mapped[int] = mapped_column(Integer)
-    updated_at_ms: Mapped[int] = mapped_column(Integer, index=True)
-    expires_at_ms: Mapped[int | None] = mapped_column(Integer, nullable=True, index=True)
+    created_at_ms: Mapped[int] = mapped_column(BigInteger)
+    updated_at_ms: Mapped[int] = mapped_column(BigInteger, index=True)
+    expires_at_ms: Mapped[int | None] = mapped_column(BigInteger, nullable=True, index=True)
 
 
 class MemoryItemRow(Base):
@@ -150,8 +152,8 @@ class MemoryItemRow(Base):
     summary: Mapped[str] = mapped_column(String)
     content: Mapped[dict[str, object]] = mapped_column(JSON_VARIANT)
     tags: Mapped[list[str]] = mapped_column(JSON_VARIANT)
-    created_at_ms: Mapped[int] = mapped_column(Integer, index=True)
-    updated_at_ms: Mapped[int] = mapped_column(Integer, index=True)
+    created_at_ms: Mapped[int] = mapped_column(BigInteger, index=True)
+    updated_at_ms: Mapped[int] = mapped_column(BigInteger, index=True)
 
 
 __all__ = [
